Add the Windows clipboard
Change notification uses the format listener rather than the viewer chain
it replaced. The chain is a linked list every watcher splices itself
into, and one badly behaved link breaks it for everyone downstream --
silently, with no error and nothing to notice. That is the likeliest
reason a shared clipboard "sometimes" stops working until something gets
restarted, and it is what Barrier still uses.

The clipboard can be held by one process at a time and applications open
it constantly, so a refusal is ordinary rather than a failure. Attempts
are repeated briefly instead of reported, which is the difference between
sharing that works and sharing that works most of the time.

Contents are promised, not supplied. Windows lets a format be offered and
produced only if something pastes it, so a large image costs nothing
until it is wanted. Formats are offered in more than one shape where
applications differ about which they ask for, and an offer this process
made itself is recognised on the way back so a clipboard is not sent
straight back to the machine it came from.

Windows' file list is translated to and from `text/uri-list`, in wide
characters so paths outside the system's code page survive. A URI naming
another machine is declined rather than turned into a local path to
nowhere.

Clippy now runs against the Windows target in CI as well. It had only
been checking compilation, and most of the platform-specific code exists
nowhere else, so its lints had been going unseen.

+++ crates/smkvm-input/src/platform/windows/mod.rs
@@ -185,7 +185,7 @@ impl Inject for WindowsInput {
             Anonymous: INPUT_0 {
                 ki: KEYBDINPUT {
                     wVk: VIRTUAL_KEY(0),
-                    wScan: (scan & 0x00FF) as u16,
+                    wScan: scan & 0x00FF,
                     dwFlags: KEYBD_EVENT_FLAGS(flags.0),
                     time: 0,
                     dwExtraInfo: INJECTED_MARKER,
